[Twinkle#67] Make project buildable

+++ twinkle-app/twinkle.bll.services/src/main/java/com/github/donkeyrit/twinkle/bll/services/DefaultCarService.java
@@ -84,7 +84,7 @@ public class DefaultCarService implements CarService {
 	}
 
 	@Override
-	public String isTaken(int carId) {
+	public String isTaken(long carId) {
 		return rentRepository.isTaken(carId) ? "lock" : "open";
 	}
 }
