+++ twinkle-app/twinkle.bll.services/src/main/java/com/github/donkeyrit/twinkle/bll/services/interfaces/CarService.java
@@ -17,5 +17,5 @@ public interface CarService {
 	Stream<CarBodyType> getCarBodyTypes();
 	int getMaxPrice();
 	PagedResultBll<Car> getPagedResult(CarQuerySpecification filter);
-	String isTaken(int carId);
+	String isTaken(long carId);
 }
