+++ twinkle-app/twinkle.dal.abstractions/src/main/java/com/github/donkeyrit/twinkle/dal/interfaces/RentRepository.java
@@ -4,5 +4,5 @@ import com.github.donkeyrit.twinkle.dal.common.repositories.GenericRepository;
 import com.github.donkeyrit.twinkle.dal.models.Rent;
 
 public interface RentRepository extends GenericRepository<Rent>  {
-	boolean isTaken(int carId);
+	boolean isTaken(long carId);
 }
